fix: job implementation

+++ src/main/java/at/ac/uibk/dps/cirrina/execution/service/ServiceImplementation.java
@@ -13,7 +13,7 @@ import java.util.concurrent.CompletableFuture;
 public abstract class ServiceImplementation {
 
   private final String name;
-  private final float cost;
+  private final double cost;
   private final boolean local;
 
   /**
@@ -23,7 +23,7 @@ public abstract class ServiceImplementation {
    * @param cost  Cost of this service implementation.
    * @param local Whether this service implementation is local.
    */
-  public ServiceImplementation(String name, float cost, boolean local) {
+  public ServiceImplementation(String name, double cost, boolean local) {
     this.name = name;
     this.cost = cost;
     this.local = local;
@@ -70,7 +70,7 @@ public abstract class ServiceImplementation {
    *
    * @return Cost.
    */
-  public float getCost() {
+  public double getCost() {
     return cost;
   }
